remove ProtectedMethodError and update CustomError and PrivateError
