+++ lib/errors/CustomError.js
@@ -1,6 +1,7 @@
 'use strict'
 
-var PrivateMethodError = require('./PrivateMethodError')
+// var PrivateMethodError = require('./PrivateMethodError')
+// var ProtectedMethodError = require('./ProtectedMethodError')
 
 module.exports = CustomError
 
@@ -8,7 +9,6 @@ CustomError.prototype = Object.create(Error.prototype)
 CustomError.prototype.constructor = CustomError
 CustomError.prototype.setMessage = setMessage
 CustomError.prototype.createStackTrace = createStackTrace
-CustomError.prototype.protect = protect
 CustomError.prototype.privatize = privatize
 
 /**
@@ -32,7 +32,6 @@ function CustomError (msg) {
  * @method
  */
 function setMessage (msg) {
-  privatize()
   if (msg && typeof msg === 'string') {
     this.message = msg
   } else if (this.constructor.name !== 'CustomError') {
@@ -49,7 +48,6 @@ function setMessage (msg) {
  * @throws {PrivateMethodError}
  */
 function createStackTrace () {
-  privatize()
   var stack = new Error().stack
   var splited = stack.split('\n')
   var modifiedStack = splited[0].concat('\n', splited.splice(3).join('\n'))
@@ -68,22 +66,8 @@ function privatize () {
   var here = getFileCall(trace, 1)
   var caller = getFileCall(trace, 3)
   if (here !== caller) {
-    throw new PrivateMethodError()
-  }
-}
-
-/**
- * Throws a ProtectedMethodError if the caller is not a child class of CustomError
- * @method
- * @returns {undefined}
- * @throws {ProtectedMethodError}
- */
-function protect (sourcePath) {
-  var trace = new Error().stack
-  var caller = getFileCall(trace, 3)
-  if (sourcePath !== caller) {
-    // throw new ProtectedMethodError()
-    throw new Error()
+    throw new Error('PrivateMethodError !!')
+    // throw new PrivateMethodError()
   }
 }
