+++ tests/errors/PrivateMethodError.spec.js
@@ -2,9 +2,11 @@
 
 var path = require('path')
 var chai = require('chai')
+var CustomError = require(path.resolve('./lib/errors/CustomError'))
 var PrivateMethodError = require(path.resolve('./lib/errors/PrivateMethodError'))
 
 var describe = global.describe
+var beforeEach = global.beforeEach
 var it = global.it
 var expect = chai.expect
 
@@ -12,7 +14,8 @@ describe('PrivateMethodError', function () {
   it('should be a function', function () {
     expect(PrivateMethodError).to.be.a('function')
   })
-  it("should has a 'DEFAULT_ERROR_MESSAGE' static constant property", function () {
+
+  it("should have a 'DEFAULT_ERROR_MESSAGE' static constant property", function () {
     expect(PrivateMethodError).to.be.a.property('DEFAULT_ERROR_MESSAGE')
   })
 
@@ -20,8 +23,20 @@ describe('PrivateMethodError', function () {
     it('should be an object', function () {
       expect(PrivateMethodError.prototype).to.be.a('object')
     })
-    it('should be an instance of Error', function () {
-      expect(PrivateMethodError.prototype).to.be.an.instanceof(Error)
+
+    it('should be an instance of CustomError', function () {
+      expect(PrivateMethodError.prototype).to.be.an.instanceof(CustomError)
+    })
+
+    describe('.constructor', function () {
+      it('should be a function', function () {
+        expect(PrivateMethodError.prototype.constructor).to.be.a('function')
+      })
+
+      it("should have a 'name' property w/ the value 'PrivateMethodError'", function () {
+        expect(PrivateMethodError.prototype.constructor).to.be.a.property('name')
+        expect(PrivateMethodError.prototype.constructor.name).to.equal('PrivateMethodError')
+      })
     })
 
     describe('.constructor()', function () {
@@ -29,51 +44,50 @@ describe('PrivateMethodError', function () {
         var err = new PrivateMethodError()
         expect(err).to.be.an.instanceof(PrivateMethodError)
       })
-      it("should has a 'name' property of type 'String'", function () {
-        var msg = 'i am private'
-        var err = new PrivateMethodError(msg)
-        expect(err).to.be.a.property('name')
-        expect(err.message).to.be.a('String')
-      })
-      it("should has a 'message' property of type 'String'", function () {
+
+      it("should have a 'message' property of type 'String'", function () {
         var msg = 'i am private'
         var err = new PrivateMethodError(msg)
         expect(err).to.be.a.property('message')
         expect(err.message).to.be.a('String')
       })
-      it("should has a 'stack' property of type 'String'", function () {
+
+      it("should have a 'stack' property of type 'String'", function () {
         var msg = 'i am private'
         var err = new PrivateMethodError(msg)
         expect(err).to.be.a.property('stack')
         expect(err.stack).to.be.a('String')
       })
-      it('should not fail if an bad type is given instead of a string as first argument', function () {
-        var err
-        var func = function () { err = new PrivateMethodError(func) }
-        var obj = function () { err = new PrivateMethodError({ foo: 'bar' }) }
-        var arr = function () { err = new PrivateMethodError([ 'foo', 'baz' ]) }
 
-        expect(func).to.not.throw('TypeError')
-        expect(err.message).to.be.a('String')
+      describe('when a non-string argument is given', function () {
+        var err0, err1, err2
+        var func, obj, arr
 
-        expect(obj).to.not.throw('TypeError')
-        expect(err.message).to.be.a('String')
+        beforeEach(function () {
+          func = function () { err0 = new PrivateMethodError(func) }
+          obj = function () { err1 = new PrivateMethodError({ foo: 'bar' }) }
+          arr = function () { err2 = new PrivateMethodError([ 'foo', 'baz' ]) }
+        })
 
-        expect(arr).to.not.throw('TypeError')
-        expect(err.message).to.be.a('String')
+        it('should not fail', function () {
+          expect(func).to.not.throw()
+          expect(obj).to.not.throw()
+          expect(arr).to.not.throw()
+        })
+
+        it('should define the "message" property as a String', function () {
+          expect(err0.message).to.be.a('String')
+          expect(err1.message).to.be.a('String')
+          expect(err2.message).to.be.a('String')
+        })
       })
-      it('should fails if we call the private createStackTrace method', function () {
-        var func = function () {
-          var err = new PrivateMethodError(func)
-          var t = err.createStackTrace()
-          console.log(t)
-        }
-        expect(func).to.throw(Error)
-      })
-      it('should has a default error message if no argument is given', function () {
-        var err = new PrivateMethodError()
-        expect(err.message).to.be.a('String')
-        expect(err.message).to.be.equal(PrivateMethodError.DEFAULT_ERROR_MESSAGE)
+
+      describe('when no argument is given', function () {
+        it('should set the default error message', function () {
+          var err = new PrivateMethodError()
+          expect(err.message).to.be.a('String')
+          expect(err.message).to.be.equal(PrivateMethodError.DEFAULT_ERROR_MESSAGE)
+        })
       })
     })
   })
